dex2oat: disable abort on soft_verifier_failures
This is not the right way to fix this issue but it works...

issue:
`dex2oat F 12-17 17:49:16 556146 556146 compiler_driver.cc:942] Had 3 soft failure(s) verifying all classes, and was asked to abort in such situations. Please check the log.
ERROR: Dex2oat failed to compile a boot image. It is likely that the boot classpath is inconsistent. Rebuild with ART_BOOT_IMAGE_EXTRA_ARGS=--runtime-arg -verbose:verifier to see verification errors.'